clean up workspace, combine utils into train
