+++ other scripts/print_trainable_params.py
